+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- FWCloud-UI user's manual static documentation access URL.
+- Config options for document root paths of FWCloud-UI and FWCloud-UI-Doc.
+
 
 ## [1.0.3] - 2021-01-25
 ### Fixed